echoファイルの分割
